src/test/java/com/adiops/apigateway/common/utils/DateUtilsTest.java
package com.adiops.apigateway.common.utils;

import static org.junit.Assert.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;

import java.time.LocalDate;
import java.util.List;

import org.junit.jupiter.api.Test;

class DateUtilsTest {

	@Test
	void test_getLastMonthDateOfCurrentYear() {
		List<LocalDate> dates=DateUtils.getLastMonthDateOfCurrentYear();
		assertNotNull(dates);
		assertEquals(LocalDate.now().getMonthValue(), dates.size());
		assertEquals(31,dates.get(0).getDayOfMonth());
		assertEquals(29,dates.get(1).getDayOfMonth());
		assertEquals(31,dates.get(2).getDayOfMonth());
		assertEquals(30,dates.get(3).getDayOfMonth());
		assertEquals(31,dates.get(4).getDayOfMonth());
		assertEquals(30,dates.get(5).getDayOfMonth());
		assertEquals(LocalDate.now().getDayOfMonth(),dates.get(6).getDayOfMonth());
	}

}
